requirements.txt
fastapi
uvicorn
python-dotenv
requests
httpx
pydantic
pinecone
python-multipart